+++ src/main/java/com/telepathicgrunt/blame/main/DispenserBlockRegistry.java
@@ -24,22 +24,22 @@ public class DispenserBlockRegistry<K, V> extends Object2ObjectOpenHashMap<K, V>
 
 	/**
 	 * For any mod to make Blame not print thousands of lines about their Dispenser Behavior registry replacement.
-	 * ONLY USE THIS IF YOUR MOD IS REPLACING 10+ BEHAVIORS. BY CONDENSING THE MESSAGES TO A SINGLE ENTRY, YOU
-	 * COULD BE HIDING INFO THAT MIGHT ACTUALLY HELP PEOPLE FIGURE OUT WHY AN ITEM DISPENSER BEHAVIOR IS BROKEN.
+	 * ONLY FOR MODS REPLACING 10+ BEHAVIORS. BY CONDENSING THE MESSAGES TO A SINGLE ENTRY, IT COULD END UP
+	 * HIDING INFO THAT MIGHT ACTUALLY HELP PEOPLE FIGURE OUT WHY AN ITEM DISPENSER BEHAVIOR IS BROKEN.
 	 *
 	 * Please be VERY detailed for summaryOfItemsAffected and reasonForBehaviorChange.
-	 * If this method is abused by other mods to hide info or for condensing less than 10 dispenser behaviors,
-	 * I will remove this exposed method in the future.
+	 * This method can easily be abused by other mods to hide info or for condensing
+	 * less than 10 dispenser behaviors which is why this method will remain not exposed.
 	 *
 	 * Remember, Blame is not supposed to be on 24/7. It is purely a diagnosis mod for weird worldgen crashes and bugs.
 	 *
-	 * @param modID The ID of your mod that wants to condense Blame's Dispenser Behavior messages about it.
+	 * @param modID The ID of the mod that wants to condense Blame's Dispenser Behavior messages about it.
 	 * @param stacktraceLineToDetect The line for Blame to look for in the stacktrace to know when to condense. Example: "vazkii.quark.content.automation.module.DispensersPlaceBlocksModule"
-	 * @param summaryOfItemsAffected Sentences describing what items your mod will be targeting to replace the behaviors of. If your mod has a config option to change what items are targeted, STATE THAT THE CONIG OPTION EXISTS HERE TOO.
-	 * @param reasonForBehaviorChange Sentences stating why your mod is replacing a ton of item's dispenser behaviors so users know what your mod is trying to do.
+	 * @param summaryOfItemsAffected Sentences describing what items the mod will be targeting to replace the behaviors of. If the mod has a config option to change what items are targeted, STATE THAT THE CONIG OPTION EXISTS HERE TOO.
+	 * @param reasonForBehaviorChange Sentences stating why the mod is replacing a ton of item's dispenser behaviors so users know what the mod is trying to do.
 	 */
-	public static void addCondensedMessage(String modID, String stacktraceLineToDetect, String summaryOfItemsAffected, String reasonForBehaviorChange){
-		messageCondenserMap.put(stacktraceLineToDetect, new MessageCondenserEntry(modID, summaryOfItemsAffected, reasonForBehaviorChange));
+	private static void addCondensedMessage(String modID, String stacktraceLineToDetect, String summaryOfItemsAffected, String reasonForBehaviorChange){
+		MESSAGE_CONDENSER_MAP.put(stacktraceLineToDetect, new MessageCondenserEntry(modID, summaryOfItemsAffected, reasonForBehaviorChange));
 	}
 
 
@@ -47,7 +47,7 @@ public class DispenserBlockRegistry<K, V> extends Object2ObjectOpenHashMap<K, V>
 
 	// Turn on registry replacement detection only after startup's putAll I do is done.
 	public Boolean startupIgnore = true;
-	private static final Map<String, MessageCondenserEntry> messageCondenserMap = new HashMap<>();
+	private static final Map<String, MessageCondenserEntry> MESSAGE_CONDENSER_MAP = new HashMap<>();
 	static{
 		// Prevent Quark's BlockBehaviour stuff from triggering Blame and printing out 7000 lines of registry replacements that they do.
 		addCondensedMessage("quark",
@@ -67,9 +67,9 @@ public class DispenserBlockRegistry<K, V> extends Object2ObjectOpenHashMap<K, V>
 			ResourceLocation itemRl = Registry.ITEM.getKey((Item)item);
 			String behaviorClassName = behavior.getClass().getName();
 
-			if(messageCondenserMap.containsKey(behaviorClassName))
+			if(MESSAGE_CONDENSER_MAP.containsKey(behaviorClassName))
 			{
-				MessageCondenserEntry entry = messageCondenserMap.get(behaviorClassName);
+				MessageCondenserEntry entry = MESSAGE_CONDENSER_MAP.get(behaviorClassName);
 				if(entry.itemBehaviorsReplaced == 0){
 					Blame.LOGGER.log(Level.ERROR, "\n****************** Blame Extra Info Report " + Blame.VERSION + " ******************" +
 							"\n   Condensed Dispenser message mode activated for " + entry.modID +