+++ CHANGELOG.md
@@ -9,8 +9,9 @@ Gotta blame the broken mods!
    (V.1.9.2 Changes) (1.16.2/1.16.3/1.16.4/1.16.5 Minecraft)
    
     Dispenser Behaviors:
-• Exposed a new method for other mods to hook into and use in order to condense Blame's Registry Replacement messages about their mods if it is excessive.
-  The method is DispenserBlockRegistry.addCondensedMessage. Mod authors, READ THE JAVADOC FOR THE METHOD CAREFULLY.
+• Added code to make adding code condensing easier in Blame. Mod Authors, PLEASE CONTACT ME if your mod register replaces 
+ dispenser behaviors for more than 10 items. There will be no exposed method as it would be too easy for mods to abuse and
+ ruin the entire point of Blame as a diagnosis tool. Instead, message condensing will be done on a case-by-case basis manually.
 
     Structures/Features:
 • Hardened error reporting so broken features/structures/registries are less likely to crash Blame before Blame can report.